% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_lm_basic_eff.R
\name{draw_lm_basic_eff}
\alias{draw_lm_basic_eff}
\title{Draws samples with a more efficient version of data agregation}
\usage{
draw_lm_basic_eff(piv, Pi, Psi, n)
}
\description{
Draws samples with a more efficient version of data agregation
}
