% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lm_basic.R
\name{est_lm_basic}
\alias{est_lm_basic}
\title{Fixed est_lm_basic}
\usage{
est_lm_basic(S, yv, k, start = 0, mod = 0, tol = 10^-8,
  maxit = 1000, out_se = FALSE, piv = NULL, Pi = NULL,
  Psi = NULL)
}
\description{
Fixed est_lm_basic
}
