% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lm_cont.R
\name{est_lm_cont}
\alias{est_lm_cont}
\title{contrainsted estimations}
\usage{
est_lm_cont(S, yv, k, start = 0, mod = 0, tol = 10^-8,
  maxit = 1000, out_se = FALSE, piv = NULL, Pi = NULL,
  Psi = NULL)
}
\description{
contrainsted estimations
}
