% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lm_basic.R
\name{find_lm_basic}
\alias{find_lm_basic}
\title{Iterates over potential start values for lm_basic}
\usage{
find_lm_basic(S, yv, k, mod = 0, nrep = 2, tol1 = 10^-5,
  tol2 = 10^-10, maxit = 1000, out_se = FALSE, piv = NULL,
  Pi = NULL, Psi = NULL)
}
\description{
Iterates over potential start values for lm_basic
}
